# Don't use unsafe string for label names in the UnpackParser