We've had [some reports](https://github.com/grafana/loki/issues/8936) that `unpack` is mangling label names. We're using `unsafeGetString` to generate these so, when the underlying byte slice changes, the label name is also changed. This change incurs an alloc to make sure the generated string wont change when the underlying bytes do.

fixes #8936