@cstyan 2023-09-07
Can we add a test case based on the two additional reports we got after https://github.com/grafana/loki/issues/8936 was originally closed?