@MasslessParticle 2023-09-07
I'd love to test this but I'm drawing a blank. I've taken user logs and they parse fine. Any thoughts on how to get this to fail a test would be appreciated.