@MasslessParticle 2023-09-07
Here's the output of benchmarks. Things appear to be a tad slower but (surprisingly?) no extra allocs:
```
goos: linux
goarch: amd64
pkg: github.com/grafana/loki/pkg/logql/log
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
                                               │  before.txt  │              after.txt              │
                                               │    sec/op    │    sec/op     vs base               │
_Parser/unpack/no_labels_hints-8                 470.2n ± 10%   476.1n ± 29%       ~ (p=0.971 n=10)
_Parser/unpack/labels_hints-8                    482.5n ±  5%   489.4n ± 18%       ~ (p=0.481 n=10)
_Parser/unpack/inline_stages-8                   503.2n ±  5%   489.1n ±  8%       ~ (p=0.739 n=10)
_Parser/unpack-not_json_line/no_labels_hints-8   14.64n ±  6%   14.12n ± 18%       ~ (p=0.382 n=10)
_Parser/unpack-not_json_line/labels_hints-8      14.54n ±  3%   13.44n ±  7%  -7.60% (p=0.005 n=10)
_Parser/unpack-not_json_line/inline_stages-8     13.18n ± 12%   13.35n ±  5%       ~ (p=0.897 n=10)
geomean                                          82.72n         81.28n        -1.74%

                                               │  before.txt  │              after.txt              │
                                               │     B/op     │    B/op     vs base                 │
_Parser/unpack/no_labels_hints-8                 80.00 ± 0%     80.00 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack/labels_hints-8                    80.00 ± 0%     80.00 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack/inline_stages-8                   80.00 ± 0%     80.00 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack-not_json_line/no_labels_hints-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack-not_json_line/labels_hints-8      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack-not_json_line/inline_stages-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                     ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                                               │  before.txt  │              after.txt              │
                                               │  allocs/op   │ allocs/op   vs base                 │
_Parser/unpack/no_labels_hints-8                 4.000 ± 0%     4.000 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack/labels_hints-8                    4.000 ± 0%     4.000 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack/inline_stages-8                   4.000 ± 0%     4.000 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack-not_json_line/no_labels_hints-8   0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack-not_json_line/labels_hints-8      0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
_Parser/unpack-not_json_line/inline_stages-8     0.000 ± 0%     0.000 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                     ²               +0.00%                ²
¹ all samples are equal
² summaries must be >0 to compute geomean
```